Final improvements and splitting pages to make code more readable
