--- src/components/PostAJobForm/style.css
@@ -1,3 +1,0 @@
-.form-wrapper {
-    width: 100%;
-}
