--- src/components/PostAProfileForm/style.css
@@ -1,3 +1,0 @@
-/* .row {
-    border: 1px solid black;        
-} */
